--- a/lib/rubocop/ast/node/mixin/collection_node.rb
+++ b/lib/rubocop/ast/node/mixin/collection_node.rb
@@ -4,13 +4,17 @@ module RuboCop
   module AST
     # A mixin that helps give collection nodes array polymorphism.
     module CollectionNode
-      extend Forwardable
-
       ARRAY_METHODS =
         (Array.instance_methods - Object.instance_methods - [:to_a]).freeze
       private_constant :ARRAY_METHODS
 
-      def_delegators :to_a, *ARRAY_METHODS
+      ARRAY_METHODS.each do |method|
+        class_eval(<<~RUBY, __FILE__, __LINE__ + 1)
+          def #{method}(...)    # def key?(...)
+            to_a.#{method}(...) #   to_a.key?(...)
+          end                   # end
+        RUBY
+      end
     end
   end
 end